Replace exit() with longjmp()
One of the very annoying thing to transfer an application to a library is exit(). The app would simply exit(), but in
library context, it would crash the whole process!

Solution: prior to call ffmpeg, set a save point(setjmp). And change all the exit to longjmp, which would return to the
save point.